// src/core/property-mock-type.enum.ts
export enum PropertyMockType {
  FirstName = 10,
  LastName = 20,
  CompanyName = 30,
  Email = 40,
  Address = 50,
  City = 60,
  State = 70,
  ZipCode = 80,
  Country = 90,

  Number = 100,
  Percentage = 110,
  Asset = 120,
  Date = 130,
  EnumValue = 140,

  FixedNUmberOfWords = 150,
  RandomNumberOfWords = 160,
  Guid = 170,
}